Fix ESP32-P4 build: remove debug traces, fix format warnings
- Removed leftover diagnostic trace (std::max type mismatch on RISC-V)
- Fixed snprintf %u -> %lu for uint32_t on RISC-V (long unsigned)
- Removed dead code (unused libavcPositions, codedIdx, ends loop)

ESP32-P4 performance results (360 MHz, 32 MB PSRAM):
  Baseline CAVLC 640x480: 50 frames in 34.4s = 1.5 fps
  High CABAC 640x480: 195 frames in 13.2s = 14.8 fps

230 desktop tests pass. ESP32-P4 builds and runs on hardware.

+++ tests/test_synthetic_quality.cpp
@@ -159,12 +159,12 @@ static double decodeAndMeasurePsnr(const char* h264Fixture, const char* rawFixtu
             {
                 char dumpPath[64];
                 std::snprintf(dumpPath, sizeof(dumpPath),
-                              "build/our_pframe%u.yuv", frameIdx);
+                              "build/our_pframe%lu.yuv", (unsigned long)frameIdx);
                 FILE* df = std::fopen(dumpPath, "wb");
                 if (!df)
                 {
                     std::snprintf(dumpPath, sizeof(dumpPath),
-                                  "../build/our_pframe%u.yuv", frameIdx);
+                                  "../build/our_pframe%lu.yuv", (unsigned long)frameIdx);
                     df = std::fopen(dumpPath, "wb");
                 }
                 if (df)
